requirements.txt
Flask==1.1.1
flask_cors==3.0.9
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
joblib==0.17.0
sklearn-pandas==2.0.2
imblearn==0.0
scikit-learn==0.23.2
pickle4==0.0.1
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.23.2
matplotlib>=1.4.3
pandas>=0.19
